+++ src/test/java/org/plumelib/util/CollectionsPlumeTest.java
@@ -173,6 +173,34 @@ final class CollectionsPlumeTest {
     assertEquals(Arrays.asList("a", "b"), c);
   }
 
+  @Test
+  void test_addIf() {
+    List<Integer> iota5 = Arrays.asList(new Integer[] {1, 2, 3, 4, 5});
+    List<Integer> odd = Arrays.asList(new Integer[] {1, 3, 5});
+    List<Integer> even = Arrays.asList(new Integer[] {2, 4});
+
+    Collection<Integer> c1 = new ArrayList<>();
+    CollectionsPlume.addIf(c1, iota5, i -> i % 2 == 1);
+    assertEquals(odd, c1);
+    Collection<Integer> c2 = new LinkedHashSet<>();
+    CollectionsPlume.addIf(c2, iota5, i -> i % 2 == 0);
+    assertEquals(even, c2);
+  }
+
+  @Test
+  void test_addIfNot() {
+    List<Integer> iota5 = Arrays.asList(new Integer[] {1, 2, 3, 4, 5});
+    List<Integer> odd = Arrays.asList(new Integer[] {1, 3, 5});
+    List<Integer> even = Arrays.asList(new Integer[] {2, 4});
+
+    Collection<Integer> c1 = new ArrayList<>();
+    CollectionsPlume.addIfNot(c1, iota5, i -> i % 2 == 1);
+    assertEquals(even, c1);
+    Collection<Integer> c2 = new LinkedHashSet<>();
+    CollectionsPlume.addIfNot(c2, iota5, i -> i % 2 == 0);
+    assertEquals(odd, c2);
+  }
+
   // public static <T> boolean hasDuplicates(List<T> a)
 
   /** Test hasNoDuplicates(). */